refined code

+++ Program.cs
@@ -14,37 +14,50 @@
                 Console.WriteLine("3) Delete Contact");
                 Console.WriteLine("4) Edit Contact");
                 Console.WriteLine("9) Exit");
-
-                int choice = int.Parse( Console.ReadLine() );
-                switch (choice)
+                string choiceString = Console.ReadLine();
+                int choice;
+                if (int.TryParse(choiceString, out choice))
                 {
-                    case 1:
-                        {
-                            addressBook.addContact();
-                            break;
-                        }
-                    case 2:
-                        {
-                            addressBook.Display();
-                            break;
-                        }
-                    case 3:
-                        {
-                            addressBook.delete();
-                            break;
-                        }
-                    case 4:
-                        {
-                            addressBook.Edit();
-                            break;
-                        }
-                    case 9:
-                        {
-                            Environment.Exit(0);
-                            break;
-                        }
+                    switch (choice)
+                    {
+                        case 1:
+                            {
+                                addressBook.addContact();
+                                break;
+                            }
+                        case 2:
+                            {
+                                addressBook.Display();
+                                break;
+                            }
+                        case 3:
+                            {
+                                addressBook.delete();
+                                break;
+                            }
+                        case 4:
+                            {
+                                addressBook.Edit();
+                                break;
+                            }
+                        case 9:
+                            {
+                                Environment.Exit(0);
+                                break;
+                            }
+                        default:
+                            {
+                                Console.WriteLine("Invalid choice.");
+                                break;
+                            }
+
+                    }
                 }
-                
+                else
+                {
+                    Console.WriteLine("Invalid choice. Please enter a valid integer choice.");
+                }
+
             }
 
         }
